-  __init__.py and gumball_machine.py: add working example of a classic state machine, wip partial impl (moved from tests)
